+++ src/components/kiosk/nyt/MovieReviews.js
@@ -2,12 +2,12 @@ import { useState, useEffect } from 'react'
 
 import axios from 'axios'
 
-import PostCardMovies from './PostCardMovies'
-import { useAuth } from '../context/AuthContext'
+import PostCardMovies from '../../PostCardMovies'
+import { useAuth } from '../../../context/AuthContext'
 
 import moment from 'moment'
 
-import data from '../data/nytMovieReviews.json'
+import data from '../../../data/nytMovieReviews.json'
 const nytMovieReviewsUrl = `https://api.nytimes.com/svc/movies/v2/reviews/picks.json?&api-key=${process.env.REACT_APP_NYT_API_KEY}`
 
 export default function MovieReviews() {