+++ src/components/kiosk/nyt/TopStories/TechTopStories.js
@@ -1,14 +1,15 @@
 import { useState, useEffect } from 'react'
+
 import axios from 'axios'
 
-import PostCard from '../PostCard'
-import { useAuth } from '../../context/AuthContext'
+import PostCard from '../../../PostCard'
+import { useAuth } from '../../../../context/AuthContext'
 
-import data from '../../data/nytArt.json'
-const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/arts.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
+import data from '../../../../data/nytTech.json'
+const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/technology.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
 /* allowed values: arts, automobiles, books, business, fashion, food, health, home, insider, magazine, movies, nyregion, obituaries, opinion, politics, realestate, science, sports, sundayreview, technology, theater, t-magazine, travel, upshot, us, world */
 
-export default function ArtTopStories() {
+export default function TechnologyTopStories() {
   const [posts, setPosts] = useState([])
   const { currentUser } = useAuth()
   useEffect(() => {